backend/requirements.txt
fastapi
uvicorn
celery
redis
pillow
